Services/kdms-ocr/requirements.txt
Flask
gunicorn
flake8
flask_restful
python-dateutil
pytesseract